# salon-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Salon API",
        "description": "API za upravljanje frizerskim salonom",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/clients": {
            "get": {
                "tags": [
                    "Korisnici"
                ],
                "summary": "Lista svih klijenata",
                "operationId": "d515553a72daec88667027f5fb84b5e5",
                "responses": {
                    "200": {
                        "description": "Lista klijenata"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/hairdressers": {
            "get": {
                "tags": [
                    "Korisnici"
                ],
                "summary": "Lista svih frizera",
                "operationId": "7457ac82385e27a61cec758a243a30a0",
                "responses": {
                    "200": {
                        "description": "Lista frizera"
                    }
                }
            },
            "post": {
                "tags": [
                    "Korisnici"
                ],
                "summary": "Kreiranje novog frizera",
                "operationId": "0c4244964d8805fce4f9443a542ce82b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Ana"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Jovic"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "ana@salon.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0601234567"
                                    },
                                    "specialization": {
                                        "type": "string",
                                        "example": "Bojenje"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "10 godina iskustva"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Frizer kreiran"
                    },
                    "422": {
                        "description": "Greška validacije"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "put": {
                "tags": [
                    "Korisnici"
                ],
                "summary": "Ažuriranje korisnika",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Ana"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Jovic"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "ana@salon.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0601234567"
                                    },
                                    "specialization": {
                                        "type": "string",
                                        "example": "Bojenje"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "10 godina iskustva"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Korisnik ažuriran"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Korisnici"
                ],
                "summary": "Brisanje korisnika",
                "operationId": "fa56cffde745d3f152f95cbacd936c0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Korisnik obrisan"
                    },
                    "404": {
                        "description": "Nije pronađen"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}